+++ lib-rt/CMakeLists.txt
@@ -5,7 +5,7 @@ find_package(Protobuf REQUIRED)
 find_package(Libdwarf REQUIRED)
 find_package(Libunwind REQUIRED)
 
-protobuf_generate_cpp (PROTO_SRCS PROTO_HDRS protobuf/chkpt.proto)
+protobuf_generate_cpp (PROTO_SRCS PROTO_HDRS protobuf/snapshot.proto)
 get_filename_component(PROTO_HDR_DIRS ${PROTO_HDRS} DIRECTORY)
 
 add_library(wanco_rt)
@@ -13,11 +13,13 @@ target_compile_options(wanco_rt PRIVATE -Wall -g)
 target_sources(wanco_rt PRIVATE
     api.cc
     wrt.cc
-    chkpt/chkpt_protobuf.cc
-    lz4/lz4.c
+    checkpoint/checkpoint.cc
+    checkpoint/stacktrace.cc
+    snapshot/snapshot_protobuf.cc
     stackmap/elf.cc
     stackmap/stackmap.cc
-    stackmap/dwarf.cc
+    stackmap/metadata.cc
+    lz4/lz4.c
     ${PROTO_SRCS}
     )
 include_directories(wanco_rt PUBLIC
