+++ lib-rt/aot.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "chkpt/chkpt.h"
+#include "snapshot/snapshot.h"
 #include "wanco.h"
 #include <cstdint>
