+++ lib-rt/checkpoint/stacktrace.h
@@ -1,0 +1,9 @@
+#pragma once
+#include "stackmap/elf.h"
+#include <vector>
+
+namespace wanco {
+
+std::vector<WasmCallStackEntry> get_stack_trace(ElfFile &elf);
+
+}
