+++ lib-rt/protobuf/snapshot.proto
