+++ lib-rt/snapshot/snapshot.h
@@ -1,5 +1,4 @@
 #pragma once
-#include "wanco.h"
 #include <deque>
 #include <fstream>
 #include <string>
