--- lib-rt/stackmap/dwarf.cc
@@ -1,6 +1,0 @@
-#include <libdwarf/libdwarf.h>
-
-namespace wanco {
-
-
-} // namespace wanco
